ui: move sumNodeStats to cluster-ui for shared use
Extract the `sumNodeStats` function and `NodeSummaryStats` type from
`db-console/src/redux/nodes.ts` into `cluster-ui/src/nodes/nodeSummaryStats.ts`
so both Redux selectors and SWR-based components can use the same
implementation. Add `draining` to `NodeSummaryStats.nodeCounts` to
match the db-console version.

The db-console `redux/nodes.ts` re-exports both symbols to keep
existing consumers working without changes.

Epic: CRDB-58145
Release note: None
