+++ pkg/ui/workspaces/cluster-ui/src/index.ts
@@ -42,11 +42,14 @@ export * from "./transactionDetails";
 export * from "./text";
 export * from "./tracez";
 export { util, api };
+export { useCluster } from "./api/clusterApi";
 export { useNodes } from "./api/nodesApi";
 export { useNodesSummary } from "./api/nodesSummaryApi";
 export { useHealth } from "./api/healthApi";
 export type { NodesSummary } from "./api/nodesSummaryApi";
 export { useConnectivity } from "./api/connectivityApi";
+export { sumNodeStats } from "./nodes/nodeSummaryStats";
+export type { NodeSummaryStats } from "./nodes/nodeSummaryStats";
 export * from "./sessions";
 export * from "./timeScaleDropdown";
 export * from "./activeExecutions";
